+++ restorers/model/nafnet/__init__.py
@@ -1,4 +1,2 @@
-from .nafblock import SimpleGate, SimplifiedChannelAttention, NAFBlock
-from .baseline import ChannelAttention, BaselineBlock
 from .nafnet import NAFNet, PixelShuffle, UpScale, BlockStack
-from .plainnet import PlainBlock
+from .plainnet import NAFBlock, SimpleGate, SimplifiedChannelAttention, ChannelAttention
