+++ restorers/tests/model/nafnet/test_nafnet.py
@@ -1,19 +1,17 @@
 import unittest
 
-import tensorflow as tf
+#import tensorflow as tf
 
-from restorers.model.nafnet import (
-    NAFBlock,
-    SimpleGate,
-    SimplifiedChannelAttention,
-    NAFNet,
-    PixelShuffle,
-    UpScale,
-    BlockStack,
-    BaselineBlock,
-    ChannelAttention,
-    PlainBlock,
-)
+# from restorers.model.nafnet import (
+#     NAFBlock,
+#     SimpleGate,
+#     SimplifiedChannelAttention,
+#     NAFNet,
+#     PixelShuffle,
+#     UpScale,
+#     BlockStack,
+#     ChannelAttention
+# )
 
 
 class NAFBlockTest(unittest.TestCase):
@@ -29,7 +27,7 @@ class BaselineBlockTest(unittest.TestCase):
     def test_baselineblock(self) -> None:
         input_shape = (1, 256, 256, 3)
         x = tf.ones(input_shape)
-        baselineblock = BaselineBlock()
+        baselineblock = NafBlock(mode='baseline')
         y = baselineblock(x)
         self.assertEqual(y.shape, x.shape)
 
@@ -38,7 +36,7 @@ class PlainBlockTest(unittest.TestCase):
     def test_plainblock(self) -> None:
         input_shape = (1, 256, 256, 3)
         x = tf.ones(input_shape)
-        plainblock = PlainBlock()
+        plainblock = NafBlock(mode='plain')
         y = plainblock(x)
         self.assertEqual(y.shape, x.shape)
 
@@ -123,14 +121,14 @@ class UpScaleTest(unittest.TestCase):
 
 class BlockStackTest(unittest.TestCase):
     def setUp(self):
-        self.block_class_list = [NAFBlock, BaselineBlock, PlainBlock]
+        self.block_class_mode = ['nafblock', 'baseline', 'plain']
         self.block_nums = [2, 3]
 
     def test_blockstack(self) -> None:
         input_shape = (1, 256, 256, 3)
         x = tf.ones(input_shape)
         for num_blocks in self.block_nums:
-            for block_class in self.block_class_list:
-                block_stack = BlockStack(block_class, num_blocks)
+            for block_class_mode in self.block_class_list:
+                block_stack = BlockStack(NAFBlock, num_blocks, mode=block_class_mode)
                 y = block_stack(x)
                 self.assertEqual(y.shape, x.shape)
